Refactor: rename into SiteCreationProgressViewModel matching class name
